docker/api-gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
prometheus-client==0.19.0
structlog==23.2.0
rich==13.7.0
typer==0.9.0
aiofiles==23.2.1
jinja2==3.1.2
pandas==2.1.3
numpy==1.24.4
pillow==10.1.0
torch==2.1.1
transformers==4.35.2
langchain==0.0.340
langchain-community==0.0.10
openai==1.3.7
anthropic==0.7.8 